Move action tests under dir
